caddyesi: Refactor middleware to switch from DataTags to DataTag
This refactoring results in using only one channel for querying the
resources and later waiting in the injecting writer for the data to come
in. In each non-coalesce case we only need now one channel for the whole
operation. Enabling coalesce produces a bit more pressure on the
middleware but not on the resource backend. With coalesce we will create
a tree of three nested goroutines.